Go: refactor
