+++ internal/common/number.go
@@ -150,3 +150,24 @@ func GetInt(value interface{}) (int, bool) {
 		return 0, false
 	}
 }
+
+// CoalesceInt returns *val if val is non-nil and positive; otherwise returns
+// defaultVal. It is useful for optional int parameters (e.g. pagination)
+// where nil or a value <= 0 means "use the default".
+func CoalesceInt(val *int, defaultVal int) int {
+	if val != nil && *val > 0 {
+		return *val
+	}
+	return defaultVal
+}
+
+// IsZeroVector reports whether every element of v is zero. An empty or nil
+// slice is considered a zero vector.
+func IsZeroVector(v []float64) bool {
+	for _, x := range v {
+		if x != 0 {
+			return false
+		}
+	}
+	return true
+}
